+++ libindy/src/api/wallet.rs
@@ -164,7 +164,9 @@ pub extern fn indy_register_wallet_storage(command_handle: i32,
 ///   "storage_credentials": optional<object> Credentials for wallet storage. Storage type defines set of supported keys.
 ///                          Can be optional if storage supports default configuration.
 ///                          For 'default' storage type should be empty.
-///   "simplified_security": optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
+///   "key_derivation_method": optional<string> algorithm to use for master key derivation:
+///                          Moderate (used by default)
+///                          Interactive - less secured but faster
 /// }
 ///
 /// #Returns
@@ -235,8 +237,12 @@ pub extern fn indy_create_wallet(command_handle: i32,
 ///       "storage_credentials": optional<object> Credentials for wallet storage. Storage type defines set of supported keys.
 ///                              Can be optional if storage supports default configuration.
 ///                              For 'default' storage type should be empty.
-///       "simplified_security": optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
-///       "rekey_simplified_security": optional<bool> Use less secured but faster crypto algorithm for new master rekey derivation (false by default).
+///       "key_derivation_method": optional<string> algorithm to use for master key derivation:
+///                             Moderate (used by default)
+///                             Interactive - less secured but faster
+///       "rekey_derivation_method": optional<string> algorithm to use for master rekey derivation:
+///                             Moderate (used by default)
+///                             Interactive - less secured but faster
 ///   }
 ///
 /// #Returns
@@ -288,7 +294,9 @@ pub extern fn indy_open_wallet(command_handle: i32,
 ///   {
 ///     "path": <string>, Path of the file that contains exported wallet content
 ///     "key": <string>, Passphrase used to derive export key
-///     "simplified_security": optional<bool> Use less secured but faster crypto algorithm for export key derivation (false by default).
+///     "key_derivation_method": optional<string> algorithm to use for export key derivation:
+///                              Moderate (used by default)
+///                             Interactive - less secured but faster
 ///   }
 ///
 /// #Returns
@@ -354,7 +362,9 @@ pub extern fn indy_export_wallet(command_handle: i32,
 ///   "storage_credentials": optional<object> Credentials for wallet storage. Storage type defines set of supported keys.
 ///                          Can be optional if storage supports default configuration.
 ///                          For 'default' storage type should be empty.
-///   "simplified_security": optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
+///   "key_derivation_method": optional<string> algorithm to use for master key derivation:
+///                            Moderate (used by default)
+///                            Interactive - less secured but faster
 /// }
 /// import_config: Import settings json.
 /// {
@@ -467,7 +477,9 @@ pub extern fn indy_close_wallet(command_handle: i32,
 ///   "storage_credentials": optional<object> Credentials for wallet storage. Storage type defines set of supported keys.
 ///                          Can be optional if storage supports default configuration.
 ///                          For 'default' storage type should be empty.
-///   "simplified_security": optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
+///   "key_derivation_method": optional<string> algorithm to use for master key derivation:
+///                            Moderate (used by default)
+///                            Interactive - less secured but faster
 /// }
 ///
 /// #Returns