+++ libindy/src/domain/wallet/mod.rs
@@ -14,18 +14,28 @@ pub struct Credentials {
     pub key: String,
     pub rekey: Option<String>,
     pub storage_credentials: Option<Value>,
-    #[serde(default)]
-    pub simplified_security: bool,
-    #[serde(default)]
-    pub rekey_simplified_security: bool
+    #[serde(default = "default_key_derivation_method")]
+    pub key_derivation_method: KeyDerivationMethod,
+    #[serde(default = "default_key_derivation_method")]
+    pub rekey_key_derivation_method: KeyDerivationMethod
+}
+
+#[derive(Debug, Serialize, Deserialize)]
+pub enum KeyDerivationMethod {
+    Moderate,
+    Interactive
+}
+
+fn default_key_derivation_method() -> KeyDerivationMethod {
+    KeyDerivationMethod::Moderate
 }
 
 #[derive(Debug, Serialize, Deserialize)]
 pub struct ExportConfig {
     pub key: String,
     pub path: String,
-    #[serde(default)]
-    pub simplified_security: bool
+    #[serde(default = "default_key_derivation_method")]
+    pub key_derivation_method: KeyDerivationMethod
 }
 
 #[derive(Debug, Serialize, Deserialize)]