+++ libindy/src/services/wallet/export_import.rs
@@ -4,6 +4,7 @@ use std::io::{Write, Read, BufWriter, BufReader};
 use std::time::{SystemTime, UNIX_EPOCH};
 use rmp_serde;
 
+use domain::wallet::KeyDerivationMethod;
 use domain::wallet::export_import::{Header, EncryptionMethod, Record};
 use errors::common::CommonError;
 use utils::crypto::hash::{hash, HASHBYTES};
@@ -13,7 +14,7 @@ use super::{WalletError, Wallet, WalletRecord};
 
 const CHUNK_SIZE: usize = 1024;
 
-pub(super) fn export(wallet: &Wallet, writer: &mut Write, passphrase: &str, version: u32, simplified_security: bool) -> Result<(), WalletError> {
+pub(super) fn export(wallet: &Wallet, writer: &mut Write, passphrase: &str, version: u32, key_derivation_method: &KeyDerivationMethod) -> Result<(), WalletError> {
     if version != 0 {
         Err(CommonError::InvalidState("Unsupported version".to_string()))?;
     }
@@ -22,14 +23,13 @@ pub(super) fn export(wallet: &Wallet, writer: &mut Write, passphrase: &str, vers
     let nonce = chacha20poly1305_ietf::gen_nonce();
     let chunk_size = CHUNK_SIZE;
 
-    let encryption_method = if simplified_security {
-        EncryptionMethod::ChaCha20Poly1305IETFWithSimplify {
+    let encryption_method = match key_derivation_method {
+        KeyDerivationMethod::Moderate => EncryptionMethod::ChaCha20Poly1305IETF {
             salt: salt[..].to_vec(),
             nonce: nonce[..].to_vec(),
             chunk_size,
-        }
-    } else {
-        EncryptionMethod::ChaCha20Poly1305IETF {
+        },
+        KeyDerivationMethod::Interactive => EncryptionMethod::ChaCha20Poly1305IETFInteractive {
             salt: salt[..].to_vec(),
             nonce: nonce[..].to_vec(),
             chunk_size,
@@ -52,7 +52,7 @@ pub(super) fn export(wallet: &Wallet, writer: &mut Write, passphrase: &str, vers
 
     // Write ecnrypted
     let mut writer = chacha20poly1305_ietf::Writer::new(writer,
-                                                        chacha20poly1305_ietf::derive_key(passphrase, &salt, simplified_security)?,
+                                                        chacha20poly1305_ietf::derive_key(passphrase, &salt, key_derivation_method)?,
                                                         nonce,
                                                         chunk_size);
 
@@ -100,18 +100,21 @@ pub(super) fn import(wallet: &Wallet, reader: &mut Read, passphrase: &str) -> Re
         Err(CommonError::InvalidStructure("Unsupported version".to_string()))?;
     }
 
-    let simplify_security = header.encryption_method.simplify_security();
+    let key_derivation_method = match header.encryption_method {
+        EncryptionMethod::ChaCha20Poly1305IETF { .. } => KeyDerivationMethod::Moderate,
+        EncryptionMethod::ChaCha20Poly1305IETFInteractive { .. } => KeyDerivationMethod::Interactive,
+    };
 
     // Reads encrypted
     let mut reader = match header.encryption_method {
-        EncryptionMethod::ChaCha20Poly1305IETF { salt, nonce, chunk_size } | EncryptionMethod::ChaCha20Poly1305IETFWithSimplify { salt, nonce, chunk_size } => {
+        EncryptionMethod::ChaCha20Poly1305IETF { salt, nonce, chunk_size } | EncryptionMethod::ChaCha20Poly1305IETFInteractive { salt, nonce, chunk_size } => {
             let salt = pwhash_argon2i13::Salt::from_slice(&salt)
                 .map_err(|err| CommonError::InvalidStructure(format!("Invalid salt: {:?}", err)))?;
 
             let nonce = chacha20poly1305_ietf::Nonce::from_slice(&nonce)
                 .map_err(|err| CommonError::InvalidStructure(format!("Invalid nonce: {:?}", err)))?;
 
-            let key = chacha20poly1305_ietf::derive_key(passphrase, &salt, simplify_security)?;
+            let key = chacha20poly1305_ietf::derive_key(passphrase, &salt, &key_derivation_method)?;
 
             chacha20poly1305_ietf::Reader::new(reader, key, nonce, chunk_size)
         }
@@ -172,7 +175,7 @@ mod tests {
         _cleanup();
 
         let mut output: Vec<u8> = Vec::new();
-        export(&_wallet1(), &mut output, _passphrase(), _version1(), false).unwrap();
+        export(&_wallet1(), &mut output, _passphrase(), _version1(), &KeyDerivationMethod::Moderate).unwrap();
 
         let wallet = _wallet2();
         _assert_is_empty(&wallet);
@@ -186,7 +189,7 @@ mod tests {
         _cleanup();
 
         let mut output: Vec<u8> = Vec::new();
-        export(&_add_2_records(_wallet1()), &mut output, _passphrase(), _version1(), false).unwrap();
+        export(&_add_2_records(_wallet1()), &mut output, _passphrase(), _version1(), &KeyDerivationMethod::Moderate).unwrap();
 
         let wallet = _wallet2();
         _assert_is_empty(&wallet);
@@ -196,11 +199,11 @@ mod tests {
     }
 
     #[test]
-    fn export_import_works_for_2_items_and_simplified_security() {
+    fn export_import_works_for_2_items_and_interactive_method() {
         _cleanup();
 
         let mut output: Vec<u8> = Vec::new();
-        export(&_add_2_records(_wallet1()), &mut output, _passphrase(), _version1(), true).unwrap();
+        export(&_add_2_records(_wallet1()), &mut output, _passphrase(), _version1(), &KeyDerivationMethod::Interactive).unwrap();
 
         let wallet = _wallet2();
         _assert_is_empty(&wallet);
@@ -214,7 +217,7 @@ mod tests {
         _cleanup();
 
         let mut output: Vec<u8> = Vec::new();
-        export(&_add_300_records(_wallet1()), &mut output, _passphrase(), _version1(), false).unwrap();
+        export(&_add_300_records(_wallet1()), &mut output, _passphrase(), _version1(), &KeyDerivationMethod::Moderate).unwrap();
 
         let wallet = _wallet2();
         _assert_is_empty(&wallet);
@@ -269,7 +272,7 @@ mod tests {
         _cleanup();
 
         let mut output: Vec<u8> = Vec::new();
-        export(&_wallet1(), &mut output, _passphrase(), _version1(), false).unwrap();
+        export(&_wallet1(), &mut output, _passphrase(), _version1(), &KeyDerivationMethod::Moderate).unwrap();
 
         // Modifying one of the bytes in the header hash
         let pos = (&mut output.as_slice()).read_u32::<LittleEndian>().unwrap() as usize + 2;
@@ -284,7 +287,7 @@ mod tests {
         _cleanup();
 
         let mut output: Vec<u8> = Vec::new();
-        export(&_add_300_records(_wallet1()), &mut output, _passphrase(), _version1(), false).unwrap();
+        export(&_add_300_records(_wallet1()), &mut output, _passphrase(), _version1(), &KeyDerivationMethod::Moderate).unwrap();
 
         // Modifying one byte in the middle of encrypted part
         let pos = output.len() / 2;
@@ -299,7 +302,7 @@ mod tests {
         _cleanup();
 
         let mut output: Vec<u8> = Vec::new();
-        export(&_add_2_records(_wallet1()), &mut output, _passphrase(), _version1(), false).unwrap();
+        export(&_add_2_records(_wallet1()), &mut output, _passphrase(), _version1(), &KeyDerivationMethod::Moderate).unwrap();
 
         output.pop().unwrap();
 
@@ -312,7 +315,7 @@ mod tests {
         _cleanup();
 
         let mut output: Vec<u8> = Vec::new();
-        export(&_add_2_records(_wallet1()), &mut output, _passphrase(), _version1(), false).unwrap();
+        export(&_add_2_records(_wallet1()), &mut output, _passphrase(), _version1(), &KeyDerivationMethod::Moderate).unwrap();
 
         output.push(10);