+++ wrappers/ios/libindy-pod/Indy/Wrapper/IndyWallet.h
@@ -63,7 +63,9 @@
    "storage_credentials": optional<object> Credentials for wallet storage. Storage type defines set of supported keys.
                           Can be optional if storage supports default configuration.
                           For 'default' storage type should be empty.
-   "simplified_security": Optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
+   "key_derivation_method": optional<string> algorithm to use for master key derivation:
+                            Moderate (used by default)
+                            Interactive - less secured but faster}
  }
  @param completion Completion callback that returns error code.
 */
@@ -101,9 +103,12 @@
    "storage_credentials": optional<object> Credentials for wallet storage. Storage type defines set of supported keys.
                           Can be optional if storage supports default configuration.
                           For 'default' storage type should be empty.
-   "simplified_security": Optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
-   "rekey_simplified_security": Optional<bool> Use less secured but faster crypto algorithm for new master rekey derivation (false by default).
- }
+   "key_derivation_method": optional<string> algorithm to use for master key derivation:
+                            Moderate (used by default)
+                            Interactive - less secured but faster}
+   "rekey_derivation_method": optional<string> algorithm to use for master rekey derivation:
+                              Moderate (used by default)
+                              Interactive - less secured but faster} }
  
  @param completion Completion callback that returns error code and created handle to opened wallet to use in methods that require wallet access.
  */
@@ -146,8 +151,9 @@
    "storage_credentials": optional<object> Credentials for wallet storage. Storage type defines set of supported keys.
                           Can be optional if storage supports default configuration.
                           For 'default' storage type should be empty.
-   "simplified_security": Optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
- }
+   "key_derivation_method": optional<string> algorithm to use for master key derivation:
+                            Moderate (used by default)
+                            Interactive - less secured but faster} }
  @param completion Completion callback that returns error code.
  */
 - (void)deleteWalletWithConfig:(NSString *)config
@@ -163,8 +169,9 @@
    {
      "path": path of the file that contains exported wallet content
      "key": passphrase used to export key
-     "simplified_security": Optional<bool> Use less secured but faster crypto algorithm for export key derivation (false by default).
-   }
+   "key_derivation_method": optional<string> algorithm to use for export key derivation:
+                            Moderate (used by default)
+                            Interactive - less secured but faster}   }
  @param completion Completion callback that returns error code.
  */
 - (void)exportWalletWithHandle:(IndyHandle)walletHandle
@@ -199,8 +206,9 @@
    "storage_credentials": optional<object> Credentials for wallet storage. Storage type defines set of supported keys.
                           Can be optional if storage supports default configuration.
                           For 'default' storage type should be empty.
-   "simplified_security": Optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
- }
+   "key_derivation_method": optional<string> algorithm to use for master key derivation:
+                            Moderate (used by default)
+                            Interactive - less secured but faster} }
  @param importConfigJson  JSON containing settings for input operation.
    {
      "path": path of the file that contains exported wallet content