+++ wrappers/java/src/main/java/org/hyperledger/indy/sdk/wallet/Wallet.java
@@ -159,7 +159,9 @@ public class Wallet extends IndyJava.API implements AutoCloseable {
 	 *   "storage_credentials": optional[{credentials json}] Credentials for wallet storage. Storage type defines set of supported keys.
 	 *                          Can be optional if storage supports default configuration.
 	 *                           For 'default' storage type should be empty.
-	 *   "simplified_security": Optional[bool] Use less secured but faster crypto algorithm for master key derivation (false by default).
+	 *   "key_derivation_method": optional[string] algorithm to use for master key derivation:
+	 *                          Moderate (used by default)
+	 *                          Interactive - less secured but faster
 	 * }
 	 * @return A future that resolves no value.
 	 * @throws IndyException Thrown if a call to the underlying SDK fails.
@@ -209,8 +211,12 @@ public class Wallet extends IndyJava.API implements AutoCloseable {
 	 *       "storage_credentials": optional[{credentiails object}] Credentials for wallet storage. Storage type defines set of supported keys.
 	 *                              Can be optional if storage supports default configuration.
 	 *                               For 'default' storage type should be empty.
-	 *   "simplified_security": Optional[bool] Use less secured but faster crypto algorithm for master key derivation (false by default). 	
-	 *   "rekey_simplified_security": Optional[bool] Use less secured but faster crypto algorithm for master rekey derivation (false by default).
+	 *   "key_derivation_method": optional[string] algorithm to use for master key derivation:
+	 *                          Moderate (used by default)
+	 *                          Interactive - less secured but faster
+	 *   "rekey_derivation_method": optional[string] algorithm to use for master rekey derivation:
+	 *                              Moderate (used by default)
+	 *                              Interactive - less secured but faster
 	 *
 	 *   }
 	 * @return A future that resolves no value.
@@ -287,7 +293,9 @@ public class Wallet extends IndyJava.API implements AutoCloseable {
 	 *       "storage_credentials": optional[{credentials json}] Credentials for wallet storage. Storage type defines set of supported keys.
 	 *                              Can be optional if storage supports default configuration.
 	 *                               For 'default' storage type should be empty.
-	 *       "simplified_security": Optional[bool] Use less secured but faster crypto algorithm for master key derivation (false by default). 	
+	 *       "key_derivation_method": optional[string] algorithm to use for master key derivation:
+	 *                                Moderate (used by default)
+	 *                                Interactive - less secured but faster
 	 *   }
 	 *                       
 	 * @return A future that resolves no value.
@@ -319,7 +327,9 @@ public class Wallet extends IndyJava.API implements AutoCloseable {
 	 *   {
 	 *     "path": "string", Path of the file that contains exported wallet content
 	 *     "key": "string", Passphrase used to derive export key
-	 *     "simplified_security": Optional[bool] Use less secured but faster crypto algorithm for export key derivation (false by default).
+	 *     "key_derivation_method": optional[string] algorithm to use for export key derivation:
+	 *                            Moderate (used by default)
+	 *                            Interactive - less secured but faster
 	 *   }
 	 * @return A future that resolves no value.
 	 * @throws IndyException Thrown if a call to the underlying SDK fails.
@@ -375,7 +385,9 @@ public class Wallet extends IndyJava.API implements AutoCloseable {
 	 *   "storage_credentials": optional[{credentials json}] Credentials for wallet storage. Storage type defines set of supported keys.
 	 *                          Can be optional if storage supports default configuration.
 	 *                          For 'default' storage type should be empty.
-	 *   "simplified_security": Optional[bool] Use less secured but faster crypto algorithm for master key derivation (false by default).
+	 *   "key_derivation_method": optional[string] algorithm to use for master key derivation:
+	 *                          Moderate (used by default)
+	 *                          Interactive - less secured but faster
 	 * }	
 	 * @param importConfigJson Import settings json.
 	 * {