+++ wrappers/python/indy/wallet.py
@@ -32,7 +32,9 @@ async def create_wallet(config: str,
        "storage_credentials": optional<object> Credentials for wallet storage. Storage type defines set of supported keys.
                               Can be optional if storage supports default configuration.
                                For 'default' storage type should be empty.
-       "simplified_security": optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
+       "key_derivation_method": optional<string> algorithm to use for master key derivation:
+                                Moderate (used by default)
+                                Interactive - less secured but faster
      }
     :return: Error code
     """
@@ -90,8 +92,12 @@ async def open_wallet(config: str,
        "storage_credentials": optional<object> Credentials for wallet storage. Storage type defines set of supported keys.
                               Can be optional if storage supports default configuration.
                               For 'default' storage type should be empty.
-       "simplified_security": optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
-       "rekey_simplified_security": Optional<bool> Use less secured but faster crypto algorithm for master rekey derivation (false by default).
+       "key_derivation_method": optional<string> algorithm to use for master key derivation:
+                                Moderate (used by default)
+                                Interactive - less secured but faster
+       "rekey_derivation_method": optional<string> algorithm to use for master rekey derivation:
+                                  Moderate (used by default)
+                                  Interactive - less secured but faster
     }
     :return: Handle to opened wallet to use in methods that require wallet access.
     """
@@ -168,7 +174,9 @@ async def delete_wallet(config: str,
        "storage_credentials": optional<object> Credentials for wallet storage. Storage type defines set of supported keys.
                               Can be optional if storage supports default configuration.
                               For 'default' storage type should be empty.
-      "simplified_security": optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
+       "key_derivation_method": optional<string> algorithm to use for master key derivation:
+                                Moderate (used by default)
+                                Interactive - less secured but faster
      }
     :return:
     """
@@ -203,7 +211,9 @@ async def export_wallet(handle: int,
        {
           "path": path of the file that contains exported wallet content
           "key": passphrase used to export key
-          "simplified_security": optional<bool> Use less secured but faster crypto algorithm for export key derivation (false by default).
+          "key_derivation_method": optional<string> algorithm for учзщке key derivation:
+                                   Moderate (used by default)
+                                   Interactive - less secured but faster
        }
     :return:
     """
@@ -257,7 +267,9 @@ async def import_wallet(config: str,
        "storage_credentials": optional<object> Credentials for wallet storage. Storage type defines set of supported keys.
                               Can be optional if storage supports default configuration.
                               For 'default' storage type should be empty.
-       "simplified_security": optional<bool> Use less secured but faster crypto algorithm for master key derivation (false by default).
+       "key_derivation_method": optional<string> algorithm to use for master key derivation:
+                                Moderate (used by default)
+                                Interactive - less secured but faster
      }
     :param import_config_json: JSON containing settings for input operationЖ {
      "path": path of the file that contains exported wallet content